refactor: Move HepMC3->Internal EDM conversion to separate alg
