+++ Examples/Algorithms/Generators/ActsExamples/Generators/EventGenerator.hpp
@@ -93,11 +93,6 @@ class EventGenerator final : public ActsExamples::IReader {
   };
 
   struct Config {
-    /// Name of the output particles collection.
-    std::string outputParticles;
-    /// Name of the output vertex collection.
-    std::string outputVertices;
-
     /// Name of the output event collection.
     std::optional<std::string> outputEvent = "hepmc3_event";
 
@@ -109,23 +104,6 @@ class EventGenerator final : public ActsExamples::IReader {
     /// If true, print the listing of the generated event. This can be very
     /// verbose
     bool printListing = false;
-
-    /// Merge primary vertices
-    bool mergePrimaries = true;
-
-    /// The spatial vertex threshold below which to consider primary vertices
-    /// candidates identical.
-    double primaryVertexSpatialThreshold = 1 * Acts::UnitConstants::nm;
-
-    /// The spatial vertex threshold below which to consider secondary vertices
-    /// candidates identical.
-    double vertexSpatialThreshold = 1 * Acts::UnitConstants::um;
-
-    /// If true, merge secondary vertices that are close to their parent vertex
-    bool mergeSecondaries = true;
-
-    /// If true, check the consistency of the generated event.
-    bool checkConsistency = false;
   };
 
   EventGenerator(const Config& cfg, Acts::Logging::Level lvl);
@@ -144,9 +122,6 @@ class EventGenerator final : public ActsExamples::IReader {
  private:
   const Acts::Logger& logger() const { return *m_logger; }
 
-  void convertHepMC3ToInternalEdm(const AlgorithmContext& ctx,
-                                  const HepMC3::GenEvent& genEvent);
-
   void handleVertex(const HepMC3::GenVertex& genVertex, SimVertex& vertex,
                     std::vector<SimVertex>& vertices,
                     std::vector<SimParticle>& particles,
@@ -155,10 +130,6 @@ class EventGenerator final : public ActsExamples::IReader {
 
   Config m_cfg;
   std::unique_ptr<const Acts::Logger> m_logger;
-
-  WriteDataHandle<SimParticleContainer> m_outputParticles{this,
-                                                          "OutputParticles"};
-  WriteDataHandle<SimVertexContainer> m_outputVertices{this, "OutputVertices"};
 
   WriteDataHandle<std::shared_ptr<HepMC3::GenEvent>> m_outputEvent{
       this, "OutputEvent"};
