+++ Examples/Python/python/acts/examples/simulation.py
@@ -333,24 +333,33 @@ def addPythia8(
     evGen = acts.examples.EventGenerator(
         level=customLogLevel(),
         generators=generators,
-        outputParticles="particles_generated",
-        outputVertices="vertices_generated",
         randomNumbers=rnd,
         outputEvent=outputEvent,
         printListing=printListing,
     )
     s.addReader(evGen)
 
-    s.addWhiteboardAlias("particles", evGen.config.outputParticles)
-    s.addWhiteboardAlias("vertices_truth", evGen.config.outputVertices)
+    hepmc3Converter = acts.examples.hepmc3.HepMC3InputConverter(
+        # level=customLogLevel(),
+        level=acts.logging.DEBUG,
+        inputEvent=evGen.config.outputEvent,
+        outputParticles="particles_generated",
+        outputVertices="vertices_generated",
+        printListing=printListing,
+    )
+    s.addAlgorithm(hepmc3Converter)
+    s.addWhiteboardAlias("particles", hepmc3Converter.config.outputParticles)
+    s.addWhiteboardAlias("vertices_truth", hepmc3Converter.config.outputVertices)
 
-    s.addWhiteboardAlias("particles_generated_selected", evGen.config.outputParticles)
+    s.addWhiteboardAlias(
+        "particles_generated_selected", hepmc3Converter.config.outputParticles
+    )
 
     if printParticles:
         s.addAlgorithm(
             acts.examples.ParticlesPrinter(
                 level=customLogLevel(),
-                inputParticles=evGen.config.outputParticles,
+                inputParticles=hepmc3Converter.config.outputParticles,
             )
         )
 
@@ -362,7 +371,7 @@ def addPythia8(
         s.addWriter(
             acts.examples.CsvParticleWriter(
                 level=customLogLevel(),
-                inputParticles=evGen.config.outputParticles,
+                inputParticles=hepmc3Converter.config.outputParticles,
                 outputDir=str(outputDirCsv),
                 outputStem="particles",
             )
@@ -376,7 +385,7 @@ def addPythia8(
         s.addWriter(
             acts.examples.RootParticleWriter(
                 level=customLogLevel(),
-                inputParticles=evGen.config.outputParticles,
+                inputParticles=hepmc3Converter.config.outputParticles,
                 filePath=str(outputDirRoot / "particles.root"),
             )
         )
@@ -384,7 +393,7 @@ def addPythia8(
         s.addWriter(
             acts.examples.RootVertexWriter(
                 level=customLogLevel(),
-                inputVertices=evGen.config.outputVertices,
+                inputVertices=hepmc3Converter.config.outputVertices,
                 filePath=str(outputDirRoot / "vertices.root"),
             )
         )
